requirements.txt
# Django
django>=1.8,<1.9.0

# Django 3rd Party
djangorestframework==3.1.1
django-storages==1.1.8
boto==2.34.0

# Images
Pillow==2.6.1

# Database
psycopg2==2.5.4

# Timezone support
pytz==2014.10

# Configuration
env-excavator==1.3.0
dj-database-url==0.3.0
django-cache-url==0.8.0
python-dotenv==0.1.0

# Monitoring
rollbar==0.9.8

# Snorse
snorse==1.0.0
